+++ src/duckdb/src/include/duckdb/function/table/arrow/enum/arrow_type_info_type.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-namespace duckdb {
-
-enum class ArrowTypeInfoType : uint8_t { LIST, STRUCT, DATE_TIME, STRING, ARRAY };
-
-} // namespace duckdb