+++ src/duckdb/ub_src_planner_filter.cpp
@@ -1,8 +1,0 @@
-#include "src/planner/filter/conjunction_filter.cpp"
-
-#include "src/planner/filter/constant_filter.cpp"
-
-#include "src/planner/filter/null_filter.cpp"
-
-#include "src/planner/filter/struct_filter.cpp"
-