+++ src/reltoaltrep.cpp
@@ -1,13 +1,21 @@
 #define __STDC_FORMAT_MACROS
 
-#include "httplib.hpp"
 #include "rapi.hpp"
 #include "typesr.hpp"
 #include "reltoaltrep.hpp"
 #include "signal.hpp"
 #include "cpp11/declarations.hpp"
 
+#include "httplib.hpp"
 #include <cinttypes>
+
+#ifdef TRUE
+#undef TRUE
+#endif
+
+#ifdef FALSE
+#undef FALSE
+#endif
 
 using namespace duckdb;
 
@@ -85,8 +93,8 @@ struct AltrepRelationWrapper {
 		return GetFromExternalPtr<AltrepRelationWrapper>(x);
 	}
 
-	AltrepRelationWrapper(duckdb::shared_ptr<Relation> rel_p, bool allow_materialization_)
-	    : allow_materialization(allow_materialization_), rel(rel_p) {
+	AltrepRelationWrapper(rel_extptr_t rel_, bool allow_materialization_)
+	    : allow_materialization(allow_materialization_), rel_eptr(rel_), rel(rel_->rel) {
 	}
 
 	bool HasQueryResult() const {
@@ -99,12 +107,17 @@ struct AltrepRelationWrapper {
 				cpp11::stop("Materialization is disabled, use collect() or as_tibble() to materialize");
 			}
 
-			auto option = Rf_GetOption(RStrings::get().materialize_sym, R_BaseEnv);
-			if (option != R_NilValue && !Rf_isNull(option) && LOGICAL_ELT(option, 0) == true) {
-				Rprintf("duckplyr: materializing, review details with duckplyr::last_rel()\n");
+			auto materialize_callback = Rf_GetOption(RStrings::get().materialize_callback_sym, R_BaseEnv);
+			if (Rf_isFunction(materialize_callback)) {
+				sexp call = Rf_lang2(materialize_callback, rel_eptr);
+				Rf_eval(call, R_BaseEnv);
 			}
 
-			last_rel = rel;
+			auto materialize_message = Rf_GetOption(RStrings::get().materialize_message_sym, R_BaseEnv);
+		  if (Rf_isLogical(materialize_message) && Rf_length(materialize_message) == 1 && LOGICAL_ELT(materialize_message, 0) == true) {
+				// Legacy
+				Rprintf("duckplyr: materializing\n");
+			}
 
 			ScopedInterruptHandler signal_handler(rel->context.GetContext());
 
@@ -142,13 +155,10 @@ struct AltrepRelationWrapper {
 
 	bool allow_materialization;
 
+	rel_extptr_t rel_eptr;
 	duckdb::shared_ptr<Relation> rel;
 	duckdb::unique_ptr<QueryResult> res;
-
-	static duckdb::shared_ptr<Relation> last_rel;
 };
-
-duckdb::shared_ptr<Relation> AltrepRelationWrapper::last_rel;
 
 struct AltrepRownamesWrapper {
 
@@ -204,7 +214,7 @@ Rboolean RelToAltrep::RownamesInspect(SEXP x, int pre, int deep, int pvec,
 	AltrepRownamesWrapper::Get(x); // make sure this is alive
 	Rprintf("DUCKDB_ALTREP_REL_ROWNAMES\n");
 	return TRUE;
-	END_CPP11_EX(FALSE)
+	END_CPP11_EX(Rboolean::FALSE)
 }
 
 Rboolean RelToAltrep::RelInspect(SEXP x, int pre, int deep, int pvec, void (*inspect_subtree)(SEXP, int, int, int)) {
@@ -213,7 +223,7 @@ Rboolean RelToAltrep::RelInspect(SEXP x, int pre, int deep, int pvec, void (*ins
 	auto &col = wrapper->rel->rel->Columns()[wrapper->column_index];
 	Rprintf("DUCKDB_ALTREP_REL_VECTOR %s (%s)\n", col.Name().c_str(), col.Type().ToString().c_str());
 	return TRUE;
-	END_CPP11_EX(FALSE)
+	END_CPP11_EX(Rboolean::FALSE)
 }
 
 // this allows us to set row names on a data frame with an int argument without calling INTPTR on it
@@ -338,17 +348,12 @@ static R_altrep_class_t LogicalTypeToAltrepType(const LogicalType &type) {
 	}
 }
 
-[[cpp11::register]] SEXP rapi_get_last_rel() {
-	auto last_rel = AltrepRelationWrapper::last_rel;
-	return sexp(make_external_prot<RelationWrapper>("duckdb_relation", R_NilValue, std::move(last_rel)));
-}
-
 [[cpp11::register]] SEXP rapi_rel_to_altrep(duckdb::rel_extptr_t rel, bool allow_materialization) {
 	D_ASSERT(rel && rel->rel);
 	auto drel = rel->rel;
 	auto ncols = drel->Columns().size();
 
-	auto relation_wrapper = make_shared_ptr<AltrepRelationWrapper>(drel, allow_materialization);
+	auto relation_wrapper = make_shared_ptr<AltrepRelationWrapper>(rel, allow_materialization);
 
 	cpp11::writable::list data_frame;
 	data_frame.reserve(ncols);
@@ -385,7 +390,7 @@ static R_altrep_class_t LogicalTypeToAltrepType(const LogicalType &type) {
 	return data_frame;
 }
 
-[[cpp11::register]] SEXP rapi_rel_from_altrep_df(SEXP df, bool strict, bool allow_materialized, bool enable_materialization) {
+[[cpp11::register]] SEXP rapi_rel_from_altrep_df(SEXP df, bool strict, bool allow_materialized) {
 	if (!Rf_inherits(df, "data.frame")) {
 		if (strict) {
 			cpp11::stop("rapi_rel_from_altrep_df: Not a data.frame");
@@ -437,12 +442,6 @@ static R_altrep_class_t LogicalTypeToAltrepType(const LogicalType &type) {
 		} else {
 			return R_NilValue;
 		}
-	}
-
-	// Side effect comes last
-	// FIXME: Add separate rapi_() function for this
-	if (enable_materialization) {
-		wrapper->rel->allow_materialization = true;
 	}
 
 	return res;